requirements.txt
django-cleanup
sorl-thumbnail
# Additional requirements go here
